+++ database/migrations/2014_10_12_100000_create_password_reset_tokens_table.php
@@ -23,6 +23,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('password_reset_tokens');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('password_reset_tokens');
+        }
     }
 };
