+++ database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -27,6 +27,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('failed_jobs');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('failed_jobs');
+        }
     }
 };
